BASIC: Add an attempt to load ArmBE when EDIT keyword is used
It may fail if not available, but so did 'EDIT'. This way if it is available at least the 'EDIT' command enters the editor.
Note: this replaces the recommendation to load from ADFS::0.$.Modules.BasicEdit in the chapter "Editing BASIC files under RISC OS 2" in the BBC BASIC reference manual (!).

Version 1.53. Tagged as 'BASIC-1_53'