BuildSys: Correction to case of Iconbar toolbox component
The component name had an uppercase 'B' which, through the use of the shared TBoxMake makefile in turn became its directory name in resourcefs.
However, with non UK territory the string collation rules would not match this and the Iconbar module would die on init.
Fixes ticket #318.

Version 5.99. Tagged as 'BuildSys-5_99'