SDCMOS: Timestamp the RPi HAL on fake HAL CMOS update
As the penny pinchers didn't fit a hardware real time clock, on a CMOS write (or desktop shutdown) the HAL CMOS is stamped with the current time. On power up, if the time is found to be 1970 (the kernel default) the last shutdown time is used instead - while this time is going to be incorrect it should at least ensure the time doesn't go backwards in the moments before NetTime manages to find a time server.

Version 0.08. Tagged as 'SDCMOS-0_08'