BootCmds: Change to the BootFX progress bar SWI call.
Version 0.06 of BootFX has a non-backwards-compatible change to the
BarUpdate SWI call. As BootCommands should be the only user right
now, that's not too bad.

Admin:

This version of BootCommands will not work correctly with versions
of BootFX older than 0.06 (the progress bar will not move).

Version 1.46. Tagged as 'BootCmds-1_46'