Disc: 1 changed component   Main (Main-2_06 -> Main-2_07) Fix !Configure on the RPi - was probably due to compiler bug.
Had some very odd behaviour on the RPi when trying to launch some
plug-ins from the Configure app (usually data aborts at address &Dxxx
with the PC containing zero). This was traced to the variable
wimpevent__handlers (which is a pointer to an array) somehow ending
up pointing to NULL (Toolbox.Libs.eventlib.c.wimpevent). Other
crashes may be from similar causes but changes in compiler flags
changed the behaviour.