TimeSetup: Use the live configure bit to determine the auto DST rule checkbox
Prior to the CMOS allocation being official this plugin kept a disc copy of the 'AutoDST' flag, but was still using it even now. Generally this is not a problem as the default is off, so on running the plugin and enabling it the two would always be in sync.
However, changing the setting with *CONFIGURE would lead the plugin to lie about its state - instead using the disc copy.
Changed to pick up the CMOS allocation and not bother saving it to disc any more.

Version 0.04. Tagged as 'TimeSetup-0_04'