Parallel: Fix printing in HAL case
ioeb_busy: saved a whole instruction in the NE case
ioeb_irqwrite: we are entered with R3=IO controller, which is not useful, so reload HardwareBase each time
ioeb_strobe: the call to the HAL corrupts the byte we're printing, so stack it either side of the call

Tested printing some geometrical shapes from a PrintOut file from Draw.
Used *COPY MyTest parallel: and
     *COPY MyTest fastparallel:
and also checked with !PICSuite that the Parallel_ SWIs allow low level access to the control register (used for bit bashing).

Version 0.78. Tagged as 'Parallel-0_78'