PinSetup: Fix buffer overflow dealing with custom backdrops in paths > 78 long
The buffer for OS_ReadArgs was not allowing for MaxFilenameSize length paths, so got a buffer overflow error & quit.
For both backdrop and pinboardoptions some slack is allowed with a trailing comma for as yet unknown switches to be added in the future.
For pinboardoptions parsing the keyword string is now a constant rather than being built up with strcat/strcpy.
Fixes ticket #331.

Version 0.20. Tagged as 'PinSetup-0_20'