ScrModes: Filter out impossible modes from a modefile super set
New function, build_a_vidclist() allows the GraphicsV Vet_Mode entry point to be called when validating a mode, so the graphics driver can decide on a mode by mode basis whether it can be achieved.
This resolves the problem where an MDF has a super set of modes that the hardware only supports a subset of.
Also
* Inlined inform_modefilechanged
* Removed redundant CFLAGS from makefile
* Made a few braces and spaces consistent with their surroundings

Tested in an IOMD ROM build on a machine with and without VRAM fitted, looking in display manager that the menus are rebuilt accordingly.
Fixes ticket #315.

Version 0.35. Tagged as 'ScrModes-0_35'