ScrnSetup: Minor ScrnSetup improvements
* Get OS_Byte/OS_File/filetypes/CMOS/OS_Module defs from header files.
* Fix ticket #312 (misuse of OS_GBPB means enumerating dirs on a non FileCore filing system could get stuck).
* Simplify the steps in the mode 'Try' sequence, now the mouse is bounded in the 'Keep/Cancel' message box (with the same 5s timeout) so that you can't click elsewhere on the main window. Previously, you could click 'Set', then the 5s would elapse and the mode revert to not the one you set.
* Window size is now compared with screensize on startup and the V scroll bar turned on/off as needed.

Note that in reviewing the plugins the minimum resolution considered to fit without scroll bars is 640x480, a special case is made of the screen setup plugin in case it starts in some mode less than that and you can't reach the 'OK' button to set it to anything else.

Version 0.25. Tagged as 'ScrnSetup-0_25'