HAL_IOMD: Fix HAL_IRQSource to sign extend 'NoInterrupt' return value
The tables of priorities are looked up with LDRB, so the NoInterrupt case returns 0xFF not -1, leading to a branch through zero in the kernel's IRQ despatch.
Interrupts.s: Reordered the priority tables to be in the same order as the IRQ status registers are inspected, saving a whole instruction, which is then used to do a final compare on r0. If only LDRSB were available...
IOMD21DMA.s: UAL syntax warning resolved

Version 0.24. Tagged as 'IOMD-0_24'