FilerAct: Fix for occasional divide-by-zero in filer copies
An else clause was being run for both the "write not complete" case but also when an error had occurred in an earlier step. The error would generally be a failure to create and hence a file size of zero. Bang.
Steps to recreate (other, more obscure and hard to catch cases exist)
 1. Open ResourceFS
 2. Copy Resources:$.Resources to Resources:$.Apps
 3. As ResourceFS is read only the reads all buffer OK but the first write (should return 'read only') gives divide-by-zero

Version 0.59. Tagged as 'FilerAct-0_59'