RemPrnSpt: Makefile recreated from fragments
Usual drill
* Rename CMHG file to be the auto generated one
* Cast away some warnings
* Swap a few header files around, use defs from central headers where available
Also, add hdr/RemotePrin which in turn will be exported and end up in <swis.h>.

Eliminated the RemPrnSptMsgs module target - if we're disc based then carry round and register our own messages (through the magic of the shared makefiles), and if we're ROM based the messages are exported to ResourceFS as usual in the resources phase.

Compiled and ran, messages were correctly registered.


Version 1.11. Tagged as 'RemPrnSpt-1_11'