PrinterManager: Fix for overwriting the filename when a USB printer is manually changed to print to file
When the connection type is changed from USB to file the USB vendor/product id is kept so that, when saved, it can be recognised again later even if the printer isn't plugged in on startup.
However, on reloading the check was assuming that the vendor/product would be zero in the manually changed connection type case, which it wasn't, and would then overwrite the filename string with the DeviceFS string & special fields.
* RunImage: comment/trace update
* Sparrow: add a check of the connection type to FNis_usb_printer
* !Run: tweaked to RMENSURE the remote printer support which includes its own messages
Fixes ticket #110.

Version 1.80. Tagged as 'Manager-1_80'