Wimp: Fix bug in Wimp_WhichIcon
When matching an icon on the iconbar the window is initially matched correctly, but the "is this the right task" filter step was trying to compare the handle with -1. The result was that you would get the first match for any iconbar icon rather than a match for *your* iconbar icon.
Unrelated: Tiling.s/Iconbar.s add a couple of comments and use symbols instead of magic numbers.
Tested with a quality 3 line BASIC program, no longer returns icon #0 all the time.

Version 5.23. Tagged as 'Wimp-5_23'