SharedSnd: Fix for zombie callback loop after killing SharedSound
When the sound system is reinitialised, but SoundChannels not yet started, SharedSound books a callback 40cs in the future to retry.
When used with an 8 bit sound system this occurs repeatedly (in the hope that a 16 bit linear handler will become available) but on RMKILL the callback was not cancelled - so although the module is dead a pair of zombie callbacks keep booking & trying, booking & trying.
* Finalise handler now cancels the 40cs ticker event
* Played safe and changed the CMP of the handler addresses (clearing V) to CLRV
* Added some asserts to the service table
* Removed some unnecessary NOP's after MOV lr, pc operations.

Tested in IOMD ROM.

Version 1.12. Tagged as 'SharedSnd-1_12'