Sound1: Bugfix to auto tuning in 8 bit soundsystem case
The SWI Sound_SampleRate shouldn't be called when on an 8 bit sound system, so don't offer autotuning in that case.
As it happens, the SWI is implemented but returns 0, so the autotune was spreading the samples to infinity - resulting in silence.
In the resources, help for *TUNING moved from its syntax string to the help string.

Tested on IOMD with 8 bit sound.

Version 1.38. Tagged as 'Sound1-1_38'