System: Makefile ammended to not give up when USERIF=None
For generating !System we are careful to not include any sprites (because these would override the ones on the user's system which may not match the default USERIF for the build), this is achieved by locally passing in USERIF=None.
However, this then stops the install rule because it expands and fails to find a directory called 'None'.
* New dummy target created for install_sprites (as opposed to lots of IfThere commands)
* !Boot no longer iconsprites since it only calls !Run which does the same
* Removed !Mk as the all target did nothing useful

Version 0.93. Tagged as 'System-0_93'