System: Readd 2nd iconsprites to !Boot
Having recently removed it, this no longer deals with the case when !System is run early in the boot sequence (to get the C library loaded for the rest of !Boot) at which time Wimp$IconTheme isn't set (unless the Wimp is in ROM).
This extra *ICONSPRITES ensures the icon is right after booting.

Version 0.95. Tagged as 'System-0_95'