ThemeDefs: Add fixup step for task manager sprite pre RISC OS 5
We want two distinct sprites, 'switcher' for unknown error boxes and the OS info box, and 'taskmanager' for the tote in the bottom right of the iconbar.
If we find an older version of TaskManager where they're both 'switcher' the theme loader patches it to make them distinct.
Them, the same set of sprite resources works on everything.
Tested on RISC OS 3.50, 3.60, 3.70, 3.71, 4.02.

Version 0.09. Tagged as 'ThemeDefs-0_09'