System: Change means of icon obliteration
Using a manual IfThere ensures that if someone opens a !System distribution (with no sprites) there's no attempt to stomp on the user's sprite.
However, IfThere is in BootCommands, and that's not loaded on RISC OS 3.50 until after !System (because it needs the C library).
Change to look for non-setted-ness of System$Dir.

Retagged as System-0_95.