DesktopBoot: Boot script fixes
When softloading the Wimp pre RISC OS 3.50 we must kill the BorderUtils module (which gets loaded as a side effect of the Toolbox being loaded when !System is run). While BorderUtils is careful not to fiddle with the workspace of Wimps it doesn't recognise this check is only made when it loads (which passes as the ROM Wimp is acceptable) but subsequent softloading is missed. Generally bad things happen when it pokes the Wimps workspace on every mode change.

PreDesktop scripts were trying to fix up malformed Boot$OSVersions, but since we know we get that right those steps are redundant.

In BootRun the use of AddToRMA hadn't spotted a subtle syntax change.
The AddToRMA utility ignored a trailing 'K' and assumed the number was always in kilobytes. AddToRMA in BootCommands requires a K (or M) suffix otherwise it assumes bytes. So for a number of years only 768 bytes have been added to the RMA, which isn't enough to softload a Wimp on RISC OS 3.10.
As RISC OS 3.10 has no free pool, the AppSize command drains the RMA dry. We push 256K back into it for any subsequent requirements (notably on RISC OS 3.10, sprites are held in the RMA, and the filer seeing some apps could trigger an annoying RMA full error message).

Disabled theme loading on RISC OS 3.10, there's really not enough free RAM to usefully load that many sprites.

Typo in RO310Hook/Boot/Desktop corrected.

Version 1.17. Tagged as 'Boot-1_17'