Wimp: Fix for missing window gadgets when no tool sprites are loaded
When 'outlinefont' is enabled, the drawicon_system function used to plot the close/back/resize etc icons tries to do so from the currently selected outline font.
This usually results in various top bit set characters inappropriately being shown.
Now, when there are no tool sprites loaded the gadgets are forced to be plotted in the system font (where the Wimp has carefully redefined the font), but the remainder of the desktop (eg. the title bar, user text) stays in the selected desktop font.

Version 5.28. Tagged as 'Wimp-5_28'