Colours: Makefile tweak
Exporting resources twice without a clean would give an error from squash, because squash sets the filetype to 'squash', but maketables just truncates the tables it makes (preserving the filetype) so on the second time around squash thinks you want to decompress what is infact a palette table.
Retagged as Colours-1_85.