HAL_IOMD: Change to SDRAM probing for 256MB Kinetic
The RAM probing for Kinetic relied on the default row/column split which doesn't quite have enough address lines for 256MB SODIMMs.
Try probing each of the 3 supported row/column splits in descending order, stopping when either no RAM is found (it's not a Kinetic) or when the address lines have no holes, whichever is sooner.
Careful not to rearrange the memory when softloaded!
Tested with a 64MB, 128MB, 256MB SODIMM in both physical ROM and softload form.

Version 0.27. Tagged as 'IOMD-0_27'