DeviceFS: Adjust buffer size for disjoint APIs
The DeviceFS API is defined in terms of actual useable buffer sizes, but BufferManager only makes use of n-1 of those requested of it (or n-4 for word bnuffers) per PRM4-88.
This change adjusts the request so the client of DeviceFS gets the useable number of bytes desired.
Changes from Colin, ref http://www.riscosopen.org/forum/forums/3/topics/1895

Version 0.70. Tagged as 'DeviceFS-0_70'