InetRes: Correction to mimetype for ZIP
Zip has a specific filetype allocated, previously it was mapped to DDC, so SparkFS would (if viewing a ZIP file on a PC share for example) retype it to A91 after opening and realising it was a ZIP. Then viewing from the PC side would show it's ",a91" suffix and hence not be recognised as a ZIP any more.

Version 5.52. Tagged as 'Internet-5_52'