Draw: Add support for background blended text and text areas
The background colour for text can now be set to "none" in the colour picker to request Draw to blend to background, rather than a fixed colour.
In the saved file 0xFFFFFFFF, or -1, is used (same as foreground colour transparent) which would render gracefully as a white background on older versions of !Draw.
For text areas the \B and \C commands can use -1 -1 -1 for the RGB components to denote transparency (3 arguments are defined as required in the text area spec, even though only 1 would really be needed).
When text in a text area has a foreground colour of 'transparent' no text is plotted, this is to be consistent with the behaviour for ordinary text.
Reordered the 'Enter' submenu to have the tools listed in the same order that they appear in the tools pane.
Moved MIN/MAX macros into common header.
Draw will only offer this functionality if an appropriate background blending Font Manager is in use, this is required as !Draw appears inside RO350Hook for machines that predate the functionality.

Version 1.24. Tagged as 'Draw-1_24'