DrawFile: Add support for background blended text and text areas
Files saved by !Draw 1.24 or later can opt to have transparent background colour to me use the blending font manager. This is the corresponding change to the renderer.
Ordinary text will not be plotted if the *foreground* colour is transparent - this seems to have been a discrepancy between !Draw and DrawFile, DrawFile used to render it in the background colour instead, imagining that this was like a shadow. The spirit of the draw file spec is that the background colour is an optional hint, however.
The state flags that end up on Service_DrawObjectRender no longer contain the DrawFile modules internal flags, only those documented.

Version 1.58. Tagged as 'DrawFile-1_58'