SCSIDriver: Add support to loader for 128k EPROMs
Contrary to the silkscreen (and logically, based on the number of pins on the chip socket) the SCSI podule supports EPROMs up to 1Mbit. This is confirmed on page 72 of the user manual.
Also, correct failure to lookup the legacy IRQ semaphore location due to not initialising R0=6.
Tested in Acorn SCSI podule on RISC OS 3.70 and 5.20.

Version 1.17. Tagged as 'SCSIDriver-1_17'