UnSqzAIF: Internationalised
For completeness mainly, though it was not true that the error number 0 english message has hidden - RMREINIT of the module would have revealed the text too.
Note, as this module also appears in !System for 3.70 the message includes both token and default text (as HWDep isn't in the global messages in 3.70).

Version 0.10. Tagged as 'UnSqzAIF-0_10'