TerritoryModule: Add support for collation with numbers
Flag bit 2 to Territory_Collate will now cause the territory module to consider digits 0-9 as though they were cardinals rather than as text.
For example, an fl ligature followed by 50 would be considered to be equal to "fl000000050" when in this mode.
Test program added to test directory, the algorithm used is based on strnatcmp().

Version 0.59. Tagged as 'Module-0_59'