ChangeFSI: Fix for unaligned load processing 48 bit image data
In mappix (2 occurrences) for 48 bit data when step24=6 ChangeFSI relied on the unaligned load rotate, this is now split out for machines that can't do that.
Don't bother looking for jcompmod in ChangeFSI$Dir - it's never been supplied there.
Line 6934: fix stack imbalance.
Rehomed a few comments with the lines to which they refer, added a few comments.
Built, but not tested.

Version 1.41. Tagged as 'ChangeFSI-1_41'