Maestro: Fix for undefined variable error
Maestro was using Wimp_ReadSysInfo 2 as a proxy for finding the current eigenfactors, but in builds supporting alpha blending this can now be non-numeric.
The CASE statement only has a palette of 4 known combinations, so "A2" failed to set the sprite block variable.
Even the sprite extension was purely numeric, the same fault would occur in (for example) EX2 EY1 mode for the same reason.

Changed to read the eigenfactors from OS_ReadModeVariable.
Tested on a Risc PC, now loads successfully.

Version 1.99. Tagged as 'Maestro-1_99'