DHCP: Suggest a hostname to the DHCP server; tweak timing
As well as providing the (optional) vendor class id string, we now also suggest an (optional) hostname if one is set in Inet$HostName. This may get overridden as the hostname is one of the parameters requested from the DHCP server, but if not the server might as well have something to call us.
Change to the discover/initreboot timing. The recent change to the total timeout was wasting time by sending packets at t=0,4,8,16,32... but as the total timeout was set to 20s it only sent 3 packets then did nothing for 8s. Now, the packets are sent at t=0,2.5,5,10 with 2.5s slack at the end for a reply,
Replaced a couple of DCI protocol constants with defines.
Tested against MS Windows Server 2012 Standard, and a 3COM router, both collect DHCP addresses OK and inspection of the allocated address table shows the name of the RISC OS client.

Version 0.26. Tagged as 'DHCP-0_26'