InetSetup: Workaround Internet module panic when gateway unset
With no gateway set this plugin would cause Internet to panic when trying to write to the routing socket.
Inspection of the logic inside Internet revealed a (dummy) netmask must be provided so it fails to route and doesn't end up trying to free a block it didn't malloc.

Tested with EtherY configured with no gateway, Internet module no longer quits.

Version 0.51. Tagged as 'Source-0_51'