Maestro: Actually load the file when the RunType alias is used
Maestro was carefully parsing the file to load in the situation that it had been run because of the RunType alias, but not loading the file.
This is because the load operation was deferred into the null poll case after Wimp_Poll, except that Maestro only uses null polls while playing.
Also shuffled some of the init into FNinitialise.

Version 2.00. Tagged as 'Maestro-2_00'