ScrnSetup: Leave the selected monitor type correct after hitting 'Set'
When the monitor type is Auto at startup (type=3) and the user selects a monitor from an MDF (type=7) when pressing 'Set' the temporary instance of the ScreenModes module is killed off (in program_exit()) which restores the monitor type back to what it was when starting (type=3). That leaves DisplayManager thinking there's an MDF loaded, but the base instance of ScreenModes thinking not.
main.c/defines.h:
Minor unrelated clean ups.
monitor.c:
When saving the settings, kill off the temporary instance, apply the setting, then create a new one - this is much like adjust clicking on cancel does.
Complete a TODO with an error message if the settings file wont open.
Messages;
New error token for the TODO.

Fixes ticket #372.

Version 0.30. Tagged as 'ScrnSetup-0_30'