BootCmds: Missing message token fix
When built with CMDHELP=None the AppSizeSyntax/AppSlotSyntax/AddToRMASyntax/IfThereSyntax tokens are not present in the messages file.
In fact, for the first 3 we don't need to manually check for trailing parameters since the CMHG definition has max-args of 1, so the OS does it for us.
For IfThereSyntax we instead manually cause an IfThere syntax error in order to get the error message (with CMDHELP=None this will come from the english text, and for internationalised versions this will be looked up for us by the kernel).
Set min-args to 1 for IfThere, this saves having to construct the *If command which we know will fail (for a start there must be no 'then'!).

Tested RAM loaded, no longer reports missing message token.

Version 1.48. Tagged as 'BootCmds-1_48'