ConfigLib: Switch round precedence of search for a default CMOS setting
This resolves a small discrepancy between the "Default CMOS" menu entry in Configure and pressing "Default" on some of the dialogues with the plugins within Configure.
Now, both use
  IF ResetCMOS is in Choices use that first
  ELSE try in the read only Hook directory for this OS (via Choices$Path)
  ELSE ask the kernel what it would do for a delete-power-on

This also allows system integrators to provide a machine specific ResetCMOS in a writeable directory (Choices$Write) to override the ROM default.
Built, but not tested.

Version 0.11. Tagged as 'Configure-0_11'