ArmBE: Fix a couple of disassembler flow analysis mistakes
In two places literal pool values had been assumed to be offsets from the start of the module, when in fact they were the size of data tables.
In turn, this meant label "data_4" split what should have been a relative BL into DCB's, so post the 32 bitting process the BL was to somewhere nonsensical.
Made use of SETV/CLRV macros in a few places.
Added a warning next to a big table that is actually a code offset table - here be dragons.

Version 1.04. Tagged as 'ArmBE-1_04'