ArmBE: Regenerate the big call table
THe keypress decoder was represented as a load of DCD's which meant moving/inserting/deleting any code anywhere in the sources would make the table out of sync and ArmBE fail in mysterious ways.
Already, 2 table entries were wrong.
Rehomed a couple of data words that got disassembled as EORCS opcodes.

Version 1.05. Tagged as 'ArmBE-1_05'