T1ToFont: Fix for stiffing machine during font conversion
The memset() in startencoding gets inlined with more recent versions of the C compiler, but inspection of the generated code shows it loses track of the structure pointer which is then used in the loops that follow, meaning ~6k of RAM gets splatted somewhere bad in memory.
Since the loops actually initialise all but one of the things in the struct anyway the memset() is a bit of a waste of time, and deleting it sidesteps the compiler fault too.

Additionally, the build fixes in T1ToFont-1_30 had missed some macros in main.h, which were instead calling the erf() function in <maths.h>.

Fixes ticket #384.
Tested with Arial Type 1 font from font pack A030 from LaTeX.

Version 1.33. Tagged as 'T1ToFont-1_33'