PatchApp: Remove ineffective check for bit 31 of AIF flags word
Application note 295 boldly declared bit 31 of AIF as "application is StrongARM compatible", however the AIF docs nor the tools were ever updated to implement this.
Hence, the quick exit when the flag is set never triggers, and 17 years on it might as well not bother.
Also - added a syntax reminder in *PATCHSTATS rather than a blank line.
Resolves ticket #310.

Version 0.08. Tagged as 'Patch-0_08'