Draw: Pop to top a draw file if already open rather than loading again
As for !Edit and the filer, when asked to open the same thing a second time, just re-present the one already in memory rather than opening a second time.
We borrow the logic from !Edit - the (uncanonicalised) filename must be the same, unmodified, with the same datestamp. If all these are true just pop the existing window (or newest of the views when multiple views exist) to the top.

Draw.c:
Change the boolean expression at line 2191. While previously correct (not wimp_MDATALOAD does include wimp_MDATAOPEN) it subsequently relied on load_file to skip wimp_MDATASAVE.
DrawFileIO.c:
Add check of load/exec/modified/filename. Defer creating a new diag until later on to avoid annoying screen flicker when disposing of the diag (similarly for error cases).
Fix confusion returning FALSE as a pointer.

Version 1.26. Tagged as 'Draw-1_26'